Resolve Compiler Warnings (#64)

+++ Source/Core/Editor/Windows/GUI_Window_ShaderEditor/GUI_Window_ShaderEditor.cpp
@@ -80,7 +80,7 @@ void Window_ShaderEditor::Draw() {
 
 
             // Set Any Viewports Shaders To 0 Who Are Using This Shader
-            for (int i = 0; i < VisualRenderer_->Viewports_.size(); i++) {
+            for (int i = 0; (long)i < (long)VisualRenderer_->Viewports_.size(); i++) {
                 if (VisualRenderer_->Viewports_[i]->ShaderIndex == LivePreviewShaderIndex_) {
                     VisualRenderer_->Viewports_[i]->ShaderIndex = 0;
                 }
